# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reco API",
        "description": "Reco - Offline-First Accounting SaaS Platform API Documentation",
        "contact": {
            "name": "Reco Support",
            "url": "https://reco.app/support",
            "email": "support@reco.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://reco.aaochaletaxi.app/v1",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/subscriptions/subscribe": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Subscribe to a plan",
                "operationId": "subscribe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plan_id",
                                    "billing_cycle"
                                ],
                                "properties": {
                                    "plan_id": {
                                        "description": "Plan ID",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "billing_cycle": {
                                        "type": "string",
                                        "example": "monthly",
                                        "enum": [
                                            "monthly",
                                            "yearly"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscribed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/subscriptions/change-plan": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Change subscription plan",
                "operationId": "changePlan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plan_id",
                                    "billing_cycle"
                                ],
                                "properties": {
                                    "plan_id": {
                                        "type": "integer"
                                    },
                                    "billing_cycle": {
                                        "type": "string",
                                        "enum": [
                                            "monthly",
                                            "yearly"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plan changed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SuccessResponse": {
                "title": "SuccessResponse",
                "description": "Standard success response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "Standard error response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "An error occurred"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: 1|abc123...",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}